% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{adaptive_evalue}
\alias{adaptive_evalue}
\title{Adaptive betting strategy for e-values}
\usage{
adaptive_evalue(t_stat, n1, n0)
}
\arguments{
\item{t_stat}{T-statistic}

\item{n1}{Treatment group size}

\item{n0}{Control group size}
}
\value{
E-value using adaptive betting
}
\description{
Adaptive betting strategy for e-values
}
