% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_blocks.R
\name{add_nodes_to_tracker}
\alias{add_nodes_to_tracker}
\title{Add nodes to tracker}
\usage{
add_nodes_to_tracker(tracker, parent_ids, depth, n_children)
}
\arguments{
\item{tracker}{Node tracking object}

\item{parent_ids}{Vector of parent node IDs}

\item{depth}{Current depth}

\item{n_children}{Number of children to create}
}
\value{
Updated tracker with new node IDs
}
\description{
Add nodes to tracker
}
\keyword{internal}
