% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottom_up.R
\name{adjust_block_tests}
\alias{adjust_block_tests}
\title{Test every block and adjust the p-values}
\usage{
adjust_block_tests(
  idat,
  bdat,
  blockid = "block",
  pfn,
  p_adj_method,
  simthresh = 20,
  sims = 1000,
  fmla = YContNorm ~ trtF | blockF,
  parallel = "multicore",
  copydts = FALSE
)
}
\arguments{
\item{idat}{Data at the unit level.}

\item{bdat}{Data at the block level.}

\item{pfn}{A function to produce pvalues --- using idat.}

\item{sims}{Number of permutations for permutation-based testing}

\item{copydts}{TRUE or FALSE. TRUE if using standalone. Could be FALSE if copied objects are being sent to this function from other functions. Basically the question is whether we want this function to change the block or individual level datasets in the main environment or return a new object.}
}
\value{
A data.table containing information about the sequence tests
}
\description{
Test every block and adjust the p-values
}
