% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meinshausen_sequential.R
\name{adjust_pvalues_level}
\alias{adjust_pvalues_level}
\title{Adjust p-values at a given level using sequential rejection}
\usage{
adjust_pvalues_level(p_values, method, alpha)
}
\arguments{
\item{p_values}{Vector of p-values at current level}

\item{method}{Adjustment method}

\item{alpha}{Error rate}
}
\value{
Vector of adjusted p-values
}
\description{
Adjust p-values at a given level using sequential rejection
}
