% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alpha_fns.R
\name{alpha_investing}
\alias{alpha_investing}
\title{Alpha adjustment function: Alpha Investing}
\usage{
alpha_investing(pval, batch, nodesize)
}
\arguments{
\item{pval}{Is a numeric vector of p-values}

\item{batch}{A character or factor or even numeric vector indicating grouping among the p-values
and the order of the p-values. The first element pval and first element of batch should be the first p-value
produced in the tree and the subsequent p-values should be produced on the children of that root (or children of the children of the root, etc.)}

\item{nodesize}{Is vector indicating the information used to create each p-value. For example, it could be the number of observations, or a weighted version.}
}
\value{
A vector of alpha values
}
\description{
These function accept a vector of p-values and vector indicating batch of p-values
It produces a vector of alpha values.
}
