% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consonance_checking.R
\name{apply_consonance_correction}
\alias{apply_consonance_correction}
\title{Apply consonance correction to test results}
\usage{
apply_consonance_correction(
  node_dat,
  tracker,
  rejection_column = "testable",
  correction_method = "propagate_up"
)
}
\arguments{
\item{node_dat}{Node data table with test results}

\item{tracker}{Node tracker object}

\item{rejection_column}{Column with rejection decisions}

\item{correction_method}{Method for correction ("propagate_up", "propagate_down", "closed_testing")}
}
\value{
Corrected node data table
}
\description{
Apply consonance correction to test results
}
