% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meinshausen_sequential.R
\name{apply_sequential_rejection_meinshausen}
\alias{apply_sequential_rejection_meinshausen}
\title{Apply Sequential Rejection Principle (Goeman-Solari enhancement)}
\usage{
apply_sequential_rejection_meinshausen(node_dat, node_tracker, alpha, method)
}
\arguments{
\item{node_dat}{Node data}

\item{node_tracker}{Node tracking structure}

\item{alpha}{Error rate}

\item{method}{P-value combination method}
}
\value{
Updated node data with sequential rejections
}
\description{
Apply Sequential Rejection Principle (Goeman-Solari enhancement)
}
