% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meinshausen_sequential.R
\name{apply_traditional_meinshausen}
\alias{apply_traditional_meinshausen}
\title{Apply Traditional Meinshausen Procedure}
\usage{
apply_traditional_meinshausen(node_dat, node_tracker, alpha, method)
}
\arguments{
\item{node_dat}{Node data}

\item{node_tracker}{Node tracking structure}

\item{alpha}{Error rate}

\item{method}{P-value combination method}
}
\value{
Updated node data with traditional Meinshausen results
}
\description{
Apply Traditional Meinshausen Procedure
}
