% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection_union_tests.R
\name{build_hierarchy_structure}
\alias{build_hierarchy_structure}
\title{Build hierarchy structure from node data and tracker}
\usage{
build_hierarchy_structure(node_dat, tracker)
}
\arguments{
\item{node_dat}{Node data table}

\item{tracker}{Node tracker object}
}
\value{
List with hierarchy information
}
\description{
Build hierarchy structure from node data and tracker
}
