% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection_union_tests.R
\name{check_iu_consistency}
\alias{check_iu_consistency}
\title{Check consistency of intersection-union test results}
\usage{
check_iu_consistency(iu_results)
}
\arguments{
\item{iu_results}{Results from intersection_union_tests}
}
\value{
Logical vector indicating consistency
}
\description{
Check consistency of intersection-union test results
}
