% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{check_stopping_rule}
\alias{check_stopping_rule}
\title{Check stopping rule}
\usage{
check_stopping_rule(wealth, e_value, alpha, stopping_rule)
}
\arguments{
\item{wealth}{Current wealth}

\item{e_value}{Current e-value}

\item{alpha}{Type I error rate}

\item{stopping_rule}{Stopping rule to apply}
}
\value{
Logical indicating whether to stop
}
\description{
Check stopping rule
}
