% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_testing.R
\name{combine_pvalues}
\alias{combine_pvalues}
\title{Combine p-values using specified method}
\usage{
combine_pvalues(p_values, method)
}
\arguments{
\item{p_values}{Vector of p-values}

\item{method}{Combination method}
}
\value{
Combined p-value
}
\description{
Combine p-values using specified method
}
