% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meinshausen_sequential.R
\name{combine_pvalues_traditional}
\alias{combine_pvalues_traditional}
\title{Traditional p-value combination (used in original Meinshausen)}
\usage{
combine_pvalues_traditional(p_values, method)
}
\arguments{
\item{p_values}{Vector of p-values to combine}

\item{method}{Combination method}
}
\value{
Combined p-value
}
\description{
Traditional p-value combination (used in original Meinshausen)
}
