% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_statistics.R
\name{combined_distances_tstat}
\alias{combined_distances_tstat}
\title{The combined distances test statistic}
\usage{
combined_distances_tstat(
  fmla = Y ~ trtF | blockF,
  distfn = fast_dists_and_trans_hybrid,
  dat
)
}
\description{
The combined distances test statistic
}
\examples{
\dontrun{
# Requires dat object to be defined in environment
# tstat_dt <- combined_distances_tstat()

splitCluster("blockF", x)
## clus <- tryCatch(KMeans_rcpp(as.matrix(x), clusters = 2, num_init = 2,
##        initializer = "optimal_init")$clusters, error = function(e) {
##    kmeans(x, centers = 2)$cluster})

## Approach 2:
# clus <- Ckmeans.1d.dp(x, k = 2)$cluster
# group <- factor(as.numeric(clus == 1))
#
}
}
