% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{compute_evalue}
\alias{compute_evalue}
\title{Compute e-value for a single test}
\usage{
compute_evalue(data, outcome, treatment, wealth_rule)
}
\arguments{
\item{data}{Block data}

\item{outcome}{Outcome variable name}

\item{treatment}{Treatment variable name}

\item{wealth_rule}{Wealth rule for betting}
}
\value{
E-value
}
\description{
Compute e-value for a single test
}
