% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_sensitivity.R
\name{compute_sensitivity_summary}
\alias{compute_sensitivity_summary}
\title{Compute summary statistics for sensitivity analysis}
\usage{
compute_sensitivity_summary(sensitivity_results, alpha)
}
\arguments{
\item{sensitivity_results}{Data frame of sensitivity results}

\item{alpha}{Type I error rate}
}
\value{
Summary statistics
}
\description{
Compute summary statistics for sensitivity analysis
}
