% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{compute_t_statistic}
\alias{compute_t_statistic}
\title{Compute t-statistic for two-sample test}
\usage{
compute_t_statistic(y, trt)
}
\arguments{
\item{y}{Outcomes}

\item{trt}{Treatment indicators (0/1)}
}
\value{
T-statistic
}
\description{
Compute t-statistic for two-sample test
}
