% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_blocks.R
\name{create_node_tracker}
\alias{create_node_tracker}
\title{Create node tracker object}
\usage{
create_node_tracker()
}
\value{
List with tracker data.table and next_id counter
}
\description{
Create node tracker object
}
\keyword{internal}
