% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{dists_and_trans}
\alias{dists_and_trans}
\title{Outcome distances and transformations}
\usage{
dists_and_trans(x)
}
\arguments{
\item{x}{Is a numeric vector (the  outcome variable)}
}
\value{
A list  for inclusion in a data.table with  distances between each unit and other units as well as some transformations of those distances
}
\description{
Outcome distances and transformations
}
