% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{distsums_raw}
\alias{distsums_raw}
\title{Transformation Helper Function: Calculate Interunit  Distances}
\usage{
distsums_raw(x)
}
\arguments{
\item{x}{A  numeric vector}
}
\value{
A numeric vector of the sums  of  centered distances from  each unit the other units
}
\description{
These functions a  vector and produce a data.frame containing  distances.
}
