% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{edisti}
\alias{edisti}
\title{Outcome e-distances between treatment arms}
\usage{
edisti(x, Z)
}
\arguments{
\item{x}{Is a numeric vector (the  outcome variable)}

\item{Z}{is a binary numeric vector or a  factor vector with  only  two values}
}
\value{
Vector with  the individual level components of the energy distances between each unit and the units in  the control condition
}
\description{
Outcome e-distances between treatment arms
}
