% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{evalue_to_pvalue}
\alias{evalue_to_pvalue}
\title{Convert e-value to p-value}
\usage{
evalue_to_pvalue(e_value)
}
\arguments{
\item{e_value}{E-value}
}
\value{
Corresponding p-value
}
\description{
Convert e-value to p-value
}
