% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consonance_checking.R
\name{extract_hierarchy_relationships}
\alias{extract_hierarchy_relationships}
\title{Extract hierarchy relationships}
\usage{
extract_hierarchy_relationships(node_dat, tracker)
}
\arguments{
\item{node_dat}{Node data table}

\item{tracker}{Node tracker object}
}
\value{
List with hierarchy relationships
}
\description{
Extract hierarchy relationships
}
