% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{fast_dists_and_trans_new_parallel}
\alias{fast_dists_and_trans_new_parallel}
\title{Outcome distances and transformations: C++ OpenMP Parallel version}
\usage{
fast_dists_and_trans_new_parallel(threads)
}
\arguments{
\item{threads}{Is an integer with the number of cores to use.}
}
\value{
A distance creation function taking x (a numeric vector, usually the outcome) and a variable Z which is not used.
}
\description{
Outcome distances and transformations: C++ OpenMP Parallel version
}
