% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consonance_checking.R
\name{find_all_descendants}
\alias{find_all_descendants}
\title{Find all descendants of a node}
\usage{
find_all_descendants(node_id, tracker, all_nodes)
}
\arguments{
\item{node_id}{Node to find descendants for}

\item{tracker}{Node tracker object}

\item{all_nodes}{Vector of all node IDs}
}
\value{
Vector of descendant node IDs
}
\description{
Find all descendants of a node
}
