% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consonance_checking.R
\name{find_consonance_violations}
\alias{find_consonance_violations}
\title{Find consonance violations}
\usage{
find_consonance_violations(node_dat, hierarchy, rejection_column)
}
\arguments{
\item{node_dat}{Node data table}

\item{hierarchy}{Hierarchy relationships}

\item{rejection_column}{Column with rejection decisions}
}
\value{
List of violations
}
\description{
Find consonance violations
}
