% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_testing.R
\name{find_containing_intersections}
\alias{find_containing_intersections}
\title{Find all intersection results containing a specific node}
\usage{
find_containing_intersections(node_id, intersection_results)
}
\arguments{
\item{node_id}{Node ID to find}

\item{intersection_results}{All intersection test results}
}
\value{
List of intersection results containing the node
}
\description{
Find all intersection results containing a specific node
}
