% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_testing.R
\name{find_leaf_descendants}
\alias{find_leaf_descendants}
\title{Find leaf descendants of a node}
\usage{
find_leaf_descendants(node_id, leaf_nodes, tracker)
}
\arguments{
\item{node_id}{Node ID}

\item{leaf_nodes}{Vector of leaf node IDs}

\item{tracker}{Node tracker object}
}
\value{
Vector of leaf node IDs that are descendants
}
\description{
Find leaf descendants of a node
}
