% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_testing.R
\name{find_leaf_nodes}
\alias{find_leaf_nodes}
\title{Find leaf nodes in the hierarchy}
\usage{
find_leaf_nodes(node_dat, tracker)
}
\arguments{
\item{node_dat}{Node data table}

\item{tracker}{Node tracker object}
}
\value{
Vector of leaf node IDs
}
\description{
Find leaf nodes in the hierarchy
}
