% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_blocks.R
\name{findBlocks}
\alias{findBlocks}
\title{Test, Split, Repeat}
\usage{
findBlocks(
  idat,
  bdat,
  blockid = "block",
  splitfn,
  pfn,
  alphafn = NULL,
  simthresh = 20,
  sims = 1000,
  maxtest = 2000,
  thealpha = 0.05,
  fmla = YContNorm ~ trtF | blockF,
  parallel = "multicore",
  copydts = FALSE,
  splitby = "hwt",
  blocksize = "hwt",
  trace = FALSE
)
}
\arguments{
\item{idat}{Data at the unit level.}

\item{bdat}{Data at the block level.}

\item{blockid}{A character name of the column in idat and bdat indicating the block.}

\item{splitfn}{A function to split the data into two pieces --- using bdat}

\item{pfn}{A function to produce pvalues --- using idat.}

\item{alphafn}{A function to adjust alpha at each step. Takes one or more p-values plus a stratum or batch indicator.}

\item{simthresh}{Below which number of total observations should the p-value functions use permutations rather than asymptotic approximations}

\item{sims}{Number of permutations for permutation-based testing}

\item{maxtest}{Maximum splits or tests to do}

\item{thealpha}{Is the error rate for a given test (for cases where alphafn is NULL, or the starting alpha for alphafn not null)}

\item{fmla}{A formula with outcome~treatment assignment  | block where treatment assignment and block must be factors.}

\item{parallel}{Should the pfn use multicore processing for permutation based testing. Default is no. But could be "snow" or "multicore" following \code{approximate} in the coin package.}

\item{copydts}{TRUE or FALSE. TRUE if using findBlocks standalone. FALSE if copied objects are being sent to findBlocks from other functions.}

\item{splitby}{A string indicating which column in bdat contains a variable to guide splitting (for example, a column with block sizes or block harmonic mean weights or a column with a covariate (or a function of covariates))}

\item{blocksize}{A string with the name of the column in bdat contains information about the size of the block (or other determinant of the power of tests within that block, such as harmonic mean weight of the block or variance of the outcome within the block.)}

\item{trace}{Logical, FALSE (default) to not print split number. TRUE prints the split number.}
}
\value{
A data.table containing information about the sequence of splitting and testing
}
\description{
Split and test.
}
