% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{fixed_evalue}
\alias{fixed_evalue}
\title{Fixed betting strategy for e-values}
\usage{
fixed_evalue(t_stat, n1, n0)
}
\arguments{
\item{t_stat}{T-statistic}

\item{n1}{Treatment group size}

\item{n0}{Control group size}
}
\value{
E-value using fixed betting
}
\description{
Fixed betting strategy for e-values
}
