% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_testing.R
\name{generate_all_intersections}
\alias{generate_all_intersections}
\title{Generate all possible intersection hypotheses}
\usage{
generate_all_intersections(leaf_nodes, tracker)
}
\arguments{
\item{leaf_nodes}{Vector of leaf node IDs}

\item{tracker}{Node tracker object}
}
\value{
List of intersection hypotheses (sets of nodes)
}
\description{
Generate all possible intersection hypotheses
}
