% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consonance_checking.R
\name{generate_corrective_actions}
\alias{generate_corrective_actions}
\title{Generate corrective actions for consonance violations}
\usage{
generate_corrective_actions(violations, node_dat, hierarchy)
}
\arguments{
\item{violations}{List of violations}

\item{node_dat}{Node data table}

\item{hierarchy}{Hierarchy structure}
}
\value{
Recommended corrective actions
}
\description{
Generate corrective actions for consonance violations
}
