% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meinshausen_sequential.R
\name{get_children_nodes}
\alias{get_children_nodes}
\title{Get children nodes from tracker}
\usage{
get_children_nodes(parent_id, node_tracker)
}
\arguments{
\item{parent_id}{Parent node ID}

\item{node_tracker}{Node tracking structure}
}
\value{
Vector of child node IDs
}
\description{
Get children nodes from tracker
}
