% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meinshausen_sequential.R
\name{get_descendant_nodes}
\alias{get_descendant_nodes}
\title{Get all descendant nodes for a given node}
\usage{
get_descendant_nodes(node_id, node_tracker, node_dat)
}
\arguments{
\item{node_id}{Node ID}

\item{node_tracker}{Node tracking structure}

\item{node_dat}{Node data}
}
\value{
Data table of descendant nodes
}
\description{
Get all descendant nodes for a given node
}
