% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_blocks.R
\name{get_parent_from_tracker}
\alias{get_parent_from_tracker}
\title{Get parent ID from tracker}
\usage{
get_parent_from_tracker(tracker, node_ids)
}
\arguments{
\item{tracker}{Node tracking object}

\item{node_ids}{Vector of node IDs to find parents for}
}
\value{
Vector of parent node IDs (0 for root)
}
\description{
Get parent ID from tracker
}
\keyword{internal}
