% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_sensitivity.R
\name{hodges_lehmann_sensitivity_bounds}
\alias{hodges_lehmann_sensitivity_bounds}
\title{Hodges-Lehmann estimator sensitivity bounds}
\usage{
hodges_lehmann_sensitivity_bounds(y1, y0, gamma)
}
\arguments{
\item{y1}{Treatment group outcomes}

\item{y0}{Control group outcomes}

\item{gamma}{Sensitivity parameter}
}
\value{
Upper and lower p-value bounds
}
\description{
Hodges-Lehmann estimator sensitivity bounds
}
