% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{kelly_evalue}
\alias{kelly_evalue}
\title{Kelly optimal betting for e-values}
\usage{
kelly_evalue(t_stat, n1, n0)
}
\arguments{
\item{t_stat}{T-statistic}

\item{n1}{Treatment group size}

\item{n0}{Control group size}
}
\value{
E-value using Kelly betting
}
\description{
Kelly optimal betting for e-values
}
