% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_adjustment.R
\name{local_bh_all_ps}
\alias{local_bh_all_ps}
\title{BH local adjustment}
\usage{
local_bh_all_ps(pvals_children, alpha = 0.05)
}
\arguments{
\item{pvals_children}{Numeric vector of child p-values.}

\item{alpha}{Numeric scalar of alpha (not used)}
}
\value{
A numeric vector of BH-adjusted p-values.
}
\description{
Performs Benjamini-Hochberg adjustment on a vector of p-values.
}
\examples{
local_bh_all_ps(c(0.01, 0.04, 0.10, 0.20))
}
