% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_adjustment.R
\name{local_min_p}
\alias{local_min_p}
\title{Unadjusted local minimal p-value}
\usage{
local_min_p(pvals_children, alpha = 0.05)
}
\arguments{
\item{pvals_children}{Numeric vector of child p-values.}

\item{alpha}{Numeric scalar of alpha}
}
\value{
A single numeric value.
}
\description{
Given \eqn{k} child p-values, returns the highest p-value below alpha;
if none are below alpha returns the smallest p-value.
}
\examples{
local_min_p(c(0.01, 0.04, 0.10, 0.20))
local_min_p(c(0.10, 0.20))
}
