% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/p_adjustment.R
\name{local_unadj_all_ps}
\alias{local_unadj_all_ps}
\title{Unadjusted local step (pass-through)}
\usage{
local_unadj_all_ps(pvals_children, alpha = 0.05)
}
\arguments{
\item{pvals_children}{Numeric vector of child p-values.}

\item{alpha}{Numeric scalar of alpha (not used)}
}
\value{
A numeric vector: the same as input.
}
\description{
Returns the input p-values unmodified.
}
\examples{
local_unadj_all_ps(c(0.01, 0.04))
}
