% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_blocks.R
\name{nodeidfn}
\alias{nodeidfn}
\title{Use hashing to make a node id}
\usage{
nodeidfn(d)
}
\arguments{
\item{d}{A vector}
}
\value{
a vector of hashes
}
\description{
This is an internal function used by find_blocks. Some (rare) designs can produce
so many nodes that we have problems with integers. So, using hashes by default.
}
