% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padj_sim_fns.R
\name{padj_test_fn}
\alias{padj_test_fn}
\title{Create treatment effect, then add tau and test using the SIUP method}
\usage{
padj_test_fn(
  idat,
  bdat,
  blockid,
  trtid = "trt",
  fmla = Y ~ trtF | blockF,
  ybase,
  prop_blocks_0,
  tau_fn,
  tau_size,
  pfn,
  afn,
  p_adj_method,
  nsims,
  ncores = 1,
  splitfn = NULL,
  covariate = NULL,
  splitby = NULL,
  thealpha = 0.05
)
}
\arguments{
\item{idat}{Individual level data}

\item{trtid}{Is the name of the treatment numeric, (0,1), variable}

\item{fmla}{is a simple formula with no blocks --- we are calculating tests within all blocks.}

\item{covariate}{is the name of a covariate to be used in created covariate dependent treatment effects}

\item{reveal_and_test_fn}{Is a function that reveals observed outcomes given the potential outcomes
produced here and does tests and adjustments (using either findBlocks for the top-down method or
adjust_block_tests for the bottom-up method, for example)}
}
\value{
A pvalue for each block
}
\description{
This function returns a function that carries with it the environment containing the arguments.
The idea is to make parallelization easier.
}
