% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_fns.R
\name{pAnova}
\alias{pAnova}
\title{P-value function: Disco}
\usage{
pAnova(
  dat,
  fmla = YContNorm ~ trtF * blockF,
  simthresh = 20,
  sims = 1000,
  parallel = "no",
  ncpu = NULL,
  index = 1,
  blocks = blockF
)
}
\arguments{
\item{dat}{An object inheriting from class data.frame}

\item{fmla}{A formula  appropriate to the function. Here it should  be something like outcome~treatment*block}

\item{sims}{Either NULL (meaning use an asymptotic reference dist) or a
number (meaning sampling from the randomization distribution implied by the
formula)}

\item{index}{is a real number in  (0,2] to transform the inter-unit distances}
}
\value{
A p-value
}
\description{
These functions accept a data frame and perhaps test specific arguments
(like whether or not the test will be asympotic or simulation based). It
produces a p-value.
}
