% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{parse_formula}
\alias{parse_formula}
\title{Parse formula to extract components}
\usage{
parse_formula(formula)
}
\arguments{
\item{formula}{Test formula}
}
\value{
List with outcome and treatment components
}
\description{
Parse formula to extract components
}
