% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_sensitivity.R
\name{parse_formula_components}
\alias{parse_formula_components}
\title{Parse formula components}
\usage{
parse_formula_components(formula)
}
\arguments{
\item{formula}{Formula object}
}
\value{
List with parsed components
}
\description{
Parse formula components
}
