% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_fns.R
\name{pCombCauchyDist}
\alias{pCombCauchyDist}
\title{P-value function: Cauchy Combined Indepence Test}
\usage{
pCombCauchyDist(
  dat,
  fmla = YcontNorm ~ trtF | blockF,
  simthresh = 20,
  sims = 1000,
  parallel = "no",
  ncpu = NULL,
  distfn = fast_dists_and_trans_nomax_hybrid
)
}
\arguments{
\item{dat}{An object inheriting from class data.frame}

\item{fmla}{A formula  appropriate to the function. Here it should  be something like outcome~treatment|block}

\item{simthresh}{is the size of the data below which we use direct permutations for p-values}

\item{sims}{Either NULL (meaning use an asymptotic reference dist) or a
number (meaning sampling from the randomization distribution implied by the
formula)}

\item{parallel}{is "no" then parallelization is not required, otherwise it
is "multicore" or "snow" in the call to \code{coin::independence_test()} (see
help for coin::approximate()). Also, if parallel is not "no" and
\code{adaptive_dist_function} is TRUE, then an openmp version of the distance
creation function is called using \code{ncpu} threads (or
\code{parallel::detectCores(logical=FALSE)} cores).}

\item{ncpu}{is number of cpus  to be used for parallel operation.}

\item{distfn}{is  a function that produces one or more vectors (a data frame
or matrix) of the  same number of  rows as the dat.  Here until further
notice users should leave it at \code{mean_dist_raw_rank_tanh} since that function is
purpose built for this function.}
}
\value{
A p-value
}
\description{
P-value function: Cauchy Combined Indepence Test
}
\details{
This  function combines the p-values from univariate tests using
using (1)  the raw outcome, (2) the rank-transformed outcome, (3) the tanh
transformed raw outcome (another test statistic that is more sensitive when
there is skew), (4) the mean difference in  raw outcome Euclidean distances
bwtween treated and control observations within block;  (5) the mean
difference in ranked outcome Euclidean distances bwtween treated and control
observations; (6) a Hotelling-T style quadratic combination of the preceding
test statistics. Inspired by Rizzo and Székely's work on Euclidean distance
based testing and by Liu and Xie (2020) on the Cauchy Combination Test and
Hansen and Bowers (2008) on omnibus tests. Distance and ranks and other
transformations are all calculated by block when block is supplied in the
formula.
}
