% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_sensitivity.R
\name{plot_design_sensitivity}
\alias{plot_design_sensitivity}
\title{Plot sensitivity analysis results}
\usage{
plot_design_sensitivity(sensitivity_analysis, blocks = NULL)
}
\arguments{
\item{sensitivity_analysis}{Results from design_sensitivity_analysis}

\item{blocks}{Vector of block IDs to plot (NULL for all)}
}
\value{
ggplot object
}
\description{
Plot sensitivity analysis results
}
