% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection_union_tests.R
\name{plot_intersection_union_results}
\alias{plot_intersection_union_results}
\title{Visualize intersection-union test results}
\usage{
plot_intersection_union_results(iu_results, highlight_inconsistencies = TRUE)
}
\arguments{
\item{iu_results}{Results from intersection_union_tests}

\item{highlight_inconsistencies}{Logical, whether to highlight inconsistencies}
}
\value{
ggplot object
}
\description{
Visualize intersection-union test results
}
