% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pval_fns.R
\name{pWilcox}
\alias{pWilcox}
\title{P-value function: Wilcox Test}
\usage{
pWilcox(
  dat,
  fmla = YContNorm ~ trtF | blockF,
  simthresh = 20,
  sims = 1000,
  parallel = "no",
  ncpu = NULL
)
}
\arguments{
\item{dat}{An object inheriting from class data.frame}

\item{fmla}{outcome~treatment factor | block factor (following \code{coin} API).}

\item{simthresh}{Below which number of total observations should the p-value
functions use permutations rather than asymptotic approximations}

\item{sims}{Either NULL (meaning use an asymptotic reference dist) or a
number (meaning sampling from the randomization distribution implied by the
formula)}

\item{parallel}{Should the function use multicore processing for permutation
based testing. Default is no. But could be "snow" or "multicore" following
\code{approximate} in the coin package.}

\item{ncpu}{is the number of workers (for "snow") or cores (for "multicore").}
}
\value{
A p-value
}
\description{
These functions accept a data frame and perhaps test specific arguments
(like whether or not the test will be asympotic or simulation based). It
produces a p-value.
}
