% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padj_sim_fns.R
\name{reveal_po_and_test}
\alias{reveal_po_and_test}
\title{Repeat experiment, reveal treatment effects from the potential outcomes, test within each block, summarize}
\usage{
reveal_po_and_test(
  idat,
  bdat,
  blockid,
  trtid,
  fmla = NULL,
  ybase,
  y1var,
  prop_blocks_0,
  tau_fn,
  tau_size,
  pfn,
  p_adj_method = "fdr",
  splitfn = NULL,
  splitby = NULL,
  thealpha = 0.05
)
}
\arguments{
\item{y1var}{Is the name of the potential outcome to treatment}

\item{p_adj_method}{Is the argument to p.adjust()}

\item{splitfn}{Must be null. Only exists so that we can use the same efffect creation and testing function for both approaches.}

\item{splitby}{Must be null. Only exists so that we can use the same efffect creation and testing function for both approaches.}
}
\value{
False positive proportion out of the tests across the blocks, The false discovery rate (proportion rejected of false nulls out of all rejections), the power of the adjusted tests across blocks (the proportion of correctly rejected hypotheses out of all correct hypotheses --- in this case correct means non-null), and power of the unadjusted test (proportion correctly rejected out of  all correct hypothesis, but using unadjusted p-values).
}
\description{
The function does hypothesis tests within blocks and then summarizes the results  of this testing across the blocks. It is designed for use by padj_test_fn.
}
