% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padj_sim_fns.R
\name{reveal_po_and_test_siup}
\alias{reveal_po_and_test_siup}
\title{Repeat experiment, reveal treatment effects from the potential outcomes, test within partitions, summarize}
\usage{
reveal_po_and_test_siup(
  idat,
  bdat,
  blockid,
  trtid,
  fmla = Y ~ newZF | blockF,
  ybase,
  y1var,
  prop_blocks_0,
  tau_fn,
  tau_size,
  pfn,
  afn,
  p_adj_method = NULL,
  copydts = FALSE,
  splitfn,
  splitby,
  thealpha = 0.05
)
}
\arguments{
\item{y1var}{Is the name of the potential outcome to treatment}

\item{p_adj_method}{Is the argument to p.adjust()}
}
\value{
False positive proportion out of the tests across the blocks, The false discovery rate (proportion rejected of false nulls out of all rejections), the power of the adjusted tests across blocks (the proportion of correctly rejected hypotheses out of all correct hypotheses --- in this case correct means non-null), and power of the unadjusted test (proportion correctly rejected out of  all correct hypothesis, but using unadjusted p-values).
}
\description{
Repeat experiment, reveal treatment effects from the potential outcomes, test within partitions, summarize
}
\details{
The function does hypothesis tests within partitions of blocks (including individual blocks depending on the splitting algorithmn) and then summarizes the results  of this testing across the blocks. It very much depends  on padj_test_fn.
}
