% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_sensitivity.R
\name{sign_test_sensitivity_bounds}
\alias{sign_test_sensitivity_bounds}
\title{Sign test sensitivity bounds}
\usage{
sign_test_sensitivity_bounds(y1, y0, gamma)
}
\arguments{
\item{y1}{Treatment group outcomes}

\item{y0}{Control group outcomes}

\item{gamma}{Sensitivity parameter}
}
\value{
Upper and lower p-value bounds
}
\description{
Sign test sensitivity bounds
}
