% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting_fns.R
\name{splitLOO}
\alias{splitLOO}
\title{Splitting function: Leave One Out}
\usage{
splitLOO(bid, x)
}
\arguments{
\item{bid}{Block id}

\item{x}{A vector that we can use to order the blocks. A block-level value.
Like N or harmonic mean weight of the block.}
}
\value{
A factor categorizing blocks into groups.
}
\description{
A splitting function takes block ids and block ordering
vector (or vectors) and produces a factor that assigns some block ids to one
group or another group.
}
\examples{
# Leave-one-out splitting - focuses on largest block vs rest
block_ids <- c("B1", "B2", "B3", "B4", "B5")
block_weights <- c(0.1, 0.3, 0.8, 0.2, 0.4)  # B3 is largest

groups <- splitLOO(block_ids, block_weights)

# Show which block is isolated (should be B3 with weight 0.8)
data.frame(block_id = block_ids, weight = block_weights, group = groups)

}
