% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting_fns.R
\name{splitSpecified}
\alias{splitSpecified}
\title{A set of pre-specified splits}
\usage{
splitSpecified(bid, x)
}
\arguments{
\item{bid}{is not used}

\item{x}{is a data.table object where each column from 1 to k further divides the blocks.
Column one should be highest level and each other column should be nested}
}
\description{
A set of pre-specified splits
}
