% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting_fns.R
\name{splitSpecifiedFactor}
\alias{splitSpecifiedFactor}
\title{A set of pre-specified splits}
\usage{
splitSpecifiedFactor(bid, x)
}
\arguments{
\item{bid}{Block id}

\item{x}{Is a a factor with levels like "state.district.school". The splits will occur from left to right depending on whether there is existing variation at that level}
}
\description{
This function does binary splits using a factor variable with dots separating the names of the subgroups. If there are more than two subgroups at any level, it makes one group from the largest subgroup and another group from the rest. If there are multiple subgroups the same size, it chooses the first subgroup by order of the levels of the factor.
}
