% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting_fns.R
\name{splitSpecifiedFactorMulti}
\alias{splitSpecifiedFactorMulti}
\title{A set of pre-specified splits}
\usage{
splitSpecifiedFactorMulti(bid, x)
}
\arguments{
\item{bid}{Block id}

\item{x}{Is a a factor with levels like "state.district.school". The splits will occur from left to right depending on whether there is existing variation at that level}
}
\description{
This function allows for more than two splits at each level
}
