% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meinshausen_sequential.R
\name{test_children_recursive_meinshausen}
\alias{test_children_recursive_meinshausen}
\title{Recursive testing of children in Meinshausen procedure}
\usage{
test_children_recursive_meinshausen(
  parent_id,
  node_dat,
  node_tracker,
  alpha,
  method
)
}
\arguments{
\item{parent_id}{Parent node ID}

\item{node_dat}{Node data}

\item{node_tracker}{Node tracking structure}

\item{alpha}{Error rate}

\item{method}{P-value combination method}
}
\value{
Updated node data
}
\description{
Recursive testing of children in Meinshausen procedure
}
