% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection_union_tests.R
\name{test_intersection_null}
\alias{test_intersection_null}
\title{Test intersection null hypothesis}
\usage{
test_intersection_null(p_values, method)
}
\arguments{
\item{p_values}{Vector of child p-values}

\item{method}{Method for combination}
}
\value{
Combined p-value for intersection null
}
\description{
Test intersection null hypothesis
}
