% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_testing.R
\name{test_intersections}
\alias{test_intersections}
\title{Test intersection hypotheses}
\usage{
test_intersections(intersections, node_dat, method, alpha)
}
\arguments{
\item{intersections}{List of intersection hypotheses}

\item{node_dat}{Node data table with p-values}

\item{method}{Method for combining p-values}

\item{alpha}{Type I error rate}
}
\value{
List of intersection test results
}
\description{
Test intersection hypotheses
}
