% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intersection_union_tests.R
\name{test_union_alternative}
\alias{test_union_alternative}
\title{Test union alternative hypothesis}
\usage{
test_union_alternative(p_values, method)
}
\arguments{
\item{p_values}{Vector of child p-values}

\item{method}{Method for combination}
}
\value{
Combined p-value for union alternative
}
\description{
Test union alternative hypothesis
}
