% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalues.R
\name{update_wealth}
\alias{update_wealth}
\title{Update wealth using e-value}
\usage{
update_wealth(current_wealth, e_value, alpha, wealth_rule)
}
\arguments{
\item{current_wealth}{Current wealth level}

\item{e_value}{New e-value}

\item{alpha}{Type I error rate}

\item{wealth_rule}{Wealth update rule}
}
\value{
Updated wealth
}
\description{
Update wealth using e-value
}
