% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closed_testing.R
\name{validate_fwer_control}
\alias{validate_fwer_control}
\title{Validate closed testing procedure maintains FWER control}
\usage{
validate_fwer_control(node_dat, alpha)
}
\arguments{
\item{node_dat}{Results from closed testing}

\item{alpha}{Nominal Type I error rate}
}
\value{
Logical indicating if FWER is properly controlled
}
\description{
Validate closed testing procedure maintains FWER control
}
