% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consonance_checking.R
\name{validate_logical_hierarchy}
\alias{validate_logical_hierarchy}
\title{Validate logical hierarchy structure}
\usage{
validate_logical_hierarchy(hierarchy, node_dat)
}
\arguments{
\item{hierarchy}{Hierarchy object}

\item{node_dat}{Node data table}
}
\value{
Logical consistency check results
}
\description{
Validate logical hierarchy structure
}
