% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bottom_up.R
\name{adjust_block_tests}
\alias{adjust_block_tests}
\title{Test every block and adjust the p-values}
\usage{
adjust_block_tests(
  idat,
  bdat,
  blockid = "block",
  pfn,
  p_adj_method,
  simthresh = 20,
  sims = 1000,
  fmla = YContNorm ~ trtF | blockF,
  parallel = "no",
  copydts = FALSE
)
}
\arguments{
\item{idat}{Data at the unit level as a data.table.}

\item{bdat}{Data at the block level as a data.table.}

\item{blockid}{A character name of the column in idat and bdat indicating the block.}

\item{pfn}{A function to produce pvalues --- using idat.}

\item{p_adj_method}{A string indicating which method \code{p.adjust} should use.}

\item{simthresh}{Below which number of total observations should the p-value functions use permutations rather than asymptotic approximations}

\item{sims}{Number of permutations for permutation-based testing when the total number of cases is below the threshold.}

\item{fmla}{A formula with outcome~treatment assignment  | block where treatment assignment and block must be factors. The tests will run within each block so that \verb{|block} part is not strictly important.}

\item{parallel}{Should the pfn use multicore processing for permutation based testing. Default is no. But could be "snow" or "multicore" following \code{approximate} in the coin package.}

\item{copydts}{TRUE or FALSE. TRUE if using standalone. Could be FALSE if copied objects are being sent to this function from other functions. Basically the question is whether we want this function to change the block or individual level datasets in the main environment or return a new object.}
}
\value{
A data.table containing information about the sequence tests
}
\description{
This function tests a hypothesis (of no effects currently) within each and
every block and then adjusts the results to control the False Discovery Rate
or Family-wise Error Rate  using the base R \code{p.adjust} function.
}
