% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padj_sim_fns.R
\name{calc_errs}
\alias{calc_errs}
\title{Calculate the error and success proportions of tests for a single iteration}
\usage{
calc_errs(
  testobj,
  truevar_name,
  trueeffect_tol = .Machine$double.eps,
  blockid = "bF",
  thealpha = 0.05
)
}
\arguments{
\item{testobj}{Is an object arising from \code{\link{findBlocks}} or \code{\link{adjust_block_tests}}. It will contain block-level results.}

\item{truevar_name}{Is a string indicating the name of the variable containing the true underlying causal effect (at the block level).}

\item{trueeffect_tol}{Is the smallest effect size below which we consider the effect to be zero (by default is it floating point zero).}

\item{blockid}{A character name of the column in idat and bdat indicating the block.}

\item{thealpha}{Is the error rate for a given test (for cases where alphafn is NULL, or the starting alpha for alphafn not null)}
}
\value{
False positive proportion out of the tests across the blocks, The
false discovery rate (proportion rejected of false nulls out of all
rejections), the power of the adjusted tests across blocks (the proportion
of correctly rejected hypotheses out of all correct hypotheses --- in this
case correct means non-null), and power of the unadjusted test (proportion
correctly rejected out of  all correct hypothesis, but using unadjusted
p-values).
}
\description{
This function takes output from \link{findBlocks} or an equivalent bottom-up testing function such as \code{adjust_block_tests}
and returns the proportions of errors made. This means that the input to findBlocks includes a column containing a true block-level effect.
Repeated uses of this function allow us to assess false discovery rates and family wise error rates among other metrics of testing success.
}
