% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dists.R
\name{dists_and_trans}
\alias{dists_and_trans}
\title{Outcome distances between treatment arms and transformations of distances}
\usage{
dists_and_trans(x, Z)
}
\arguments{
\item{x}{Is a numeric vector (the  outcome variable)}

\item{Z}{is just a placeholder and not used but is a part of the api}
}
\value{
A list  for inclusion in a data.table with  distances between each unit and other units as well as some transformations of those distances
}
\description{
Outcome distances between treatment arms and transformations of distances
}
