% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation_functions.R
\name{errsimfn}
\alias{errsimfn}
\title{Error assessment function}
\usage{
errsimfn(
  idat,
  bdat,
  pfn,
  splitfn,
  fmla = Y ~ newtrtF | blockF,
  blockid,
  trtvar,
  ybase,
  afn = NULL,
  thealpha = 0.05,
  tau_fn,
  tau_size = 0,
  prop_blocks_0 = 0,
  sims
)
}
\arguments{
\item{idat}{Unit-level data. An object inheriting from class data.table}

\item{bdat}{Block-level data. An object inheriting from class data.table}

\item{pfn}{A function to  produce p-values (see \link{pWilcox} for example).}

\item{splitfn}{A function for splitting the data, see \link{splitLOO} for example.}

\item{fmla}{A formula relating outcomes  to  treatment and blocks (to  be passed to  pfn).}

\item{thealpha}{The alpha level of the test.}

\item{tau}{A treatment effect: either a scale constant additive effect or a vector with a treatment effect to be
added to each row in idat in blocks where the effect is not  zero.}

\item{propblocks}{The proportion of blocks having any treatment effect at all.}
}
\value{
A data.table with final pvalues, the associated blocks,  the true treatment effects, and the order in which the tests were conducted.
}
\description{
This function produces p-values evaluating hypotheses of known forms to help
assess the operating characteristics of testing procedures. It creates causal effects of known forms and calls the
findBlocks function and passes objects and arguments onto it.
}
