% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{make_graph}
\alias{make_graph}
\title{Make a plot of the nodes}
\usage{
make_graph(res_graph)
}
\arguments{
\item{res_graph}{A tidygraph object produced from make_tree}
}
\value{
A ggraph object
}
\description{
Given the results of the splitting and testing algorithm in the form of a
graph from \link{make_tree}, make a node level data set for use in reporting results in terms of a binary tree graph. This does not print or plot the graph. You'll need to do that with the resulting object.
}
