% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{make_results_ggraph}
\alias{make_results_ggraph}
\title{Make a plot of the nodes}
\usage{
make_results_ggraph(res_graph)
}
\arguments{
\item{res_graph}{A tidygraph object produced from make_results_tree}
}
\value{
A ggraph object
}
\description{
Given the results of the splitting and testing algorithm in the form of a
graph from \link{make_results_tree}, make a node level data set for use in
reporting results in terms of a binary tree graph. This does not print or
plot the graph. You'll need to do that with the resulting object.
}
