% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{make_results_tree}
\alias{make_results_tree}
\title{Make a node level tree object of the results of nested testing}
\usage{
make_results_tree(orig_res, blockid = "bF", node_label = NULL)
}
\arguments{
\item{orig_res}{a results data.table output from the \code{\link{findBlocks}} function.}

\item{blockid}{is a character name for the variable containing the block id information}

\item{node_label}{is a character name for a variable containing a descriptive label for the blocks.}
}
\value{
A tbl_graph and igraph object with nodes and edges
}
\description{
Given the results of the splitting and testing algorithm, make a node level
data set for use in reporting results and as input to ggraph for
visualization in terms of a tree graph.
}
