% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{make_results_tree}
\alias{make_results_tree}
\title{Make a node-level dataset from a block-level dataset}
\usage{
make_results_tree(
  orig_res,
  block_id,
  node_label = NULL,
  return_what = "all",
  truevar_name = NULL
)
}
\arguments{
\item{orig_res}{data.table from find_blocks(); must include elements such as
\itemize{
\item node_id    (node identifier, may be factorized),
\item p1,p2,… and a pfinal*,
\item alpha1, alpha2, …
}}

\item{block_id}{name of your block ID column (e.g. "bF")}

\item{node_label}{optional name of a descriptive label column}

\item{return_what}{a character vector containing "all", "graph" (a tbl_graph
object with nodes and edges), "nodes" (a data.table with node level
information), "test_summary" (a data.table object with one row indicating
false and true discoveries, etc.)}

\item{truevar_name}{the optional name of a column recording the true
treatment effect (used here to find blocks where the true effect is 0 or
not). In some simulations we have a column called nonnull which is TRUE if
that block or node has a non-zero effect and FALSE if the block or node has
a truly zero effect. So, truevar_name can be "nonnull"}
}
\value{
a list that can contain nodes, a tbl_graph object, and/or a test_summary
}
\description{
Given the results of the splitting and testing algorithm, make a node level
data set for use in reporting results and as input to ggraph for
visualization in terms of a tree graph.
}
