% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{make_tree}
\alias{make_tree}
\title{Make a node level binary tree object}
\usage{
make_tree(orig_res, blockid = "bF")
}
\arguments{
\item{orig_res}{results data.table output from the \code{\link{findBlocks}} function.}

\item{blockdis}{Is a character name for the variable containing the block id information}
}
\value{
A tbl_graph and igraph object with nodes and edges
}
\description{
Given the results of the splitting and testing algorithm, make a node level data set for use in reporting results in terms of a binary tree graph.
}
