% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_blocks.R
\name{nodeidfn}
\alias{nodeidfn}
\title{Use hashing to make a node id (DEPRECATED)}
\usage{
nodeidfn(d)
}
\arguments{
\item{d}{A vector}
}
\value{
a vector of hashes
}
\description{
This function is deprecated in favor of numeric node tracking.
Kept for backward compatibility only.
}
