% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reporting.R
\name{report_detections}
\alias{report_detections}
\title{Return detected blocks plus info}
\usage{
report_detections(
  orig_res,
  fwer = TRUE,
  alpha = 0.05,
  only_hits = FALSE,
  autofwer = TRUE,
  blockid = "blockF"
)
}
\arguments{
\item{orig_res}{results data.table output from the \code{\link{findBlocks}} function.}

\item{fwer}{(default is TRUE) means that a block is detected (or not) using the maximum p-value associated with the
block (or the groups containing that block). fwer=FALSE to detect blocks (or groups of blocks) using FDR control.}

\item{alpha}{Is the false positive rate used for detecting an effect if it is constant (i.e. not an FDR-style approach).}

\item{only_hits}{(default FALSE) returns only the detected blocks}

\item{autofwer}{If fwer=TRUE but alpha varies, return the fdr based report.}

\item{blockid}{Name of block variable (the blocking variable is a factor)}
}
\value{
A data.table adding a column \code{hit} to the \code{res} data.table indicating a "hit" or detection for that block (or group of blocks)
}
\description{
Given the results of the splitting and testing algorithm, report on the blocks
where the null of no effects could be rejected at level alpha.
}
