% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padj_sim_fns.R
\name{reveal_po_and_test_siup}
\alias{reveal_po_and_test_siup}
\title{Repeat experiment, reveal treatment effects from the potential outcomes, test within partitions, summarize}
\usage{
reveal_po_and_test_siup(
  idat,
  bdat,
  blockid,
  trtid,
  fmla = Y ~ newZF | blockF,
  ybase,
  y1var,
  prop_blocks_0,
  tau_fn,
  tau_size,
  pfn,
  afn,
  p_adj_method = "split",
  copydts = FALSE,
  splitfn,
  splitby,
  thealpha = 0.05
)
}
\arguments{
\item{idat}{Data at the unit level.}

\item{bdat}{Data at the block level.}

\item{blockid}{A character name of the column in idat and bdat indicating the block.}

\item{trtid}{Is the name of the treatment numeric, (0,1), variable}

\item{fmla}{A formula with outcome~treatment assignment  | block where treatment assignment and block must be factors.}

\item{ybase}{Is the potential outcome to control upon which the treatment effect will be built}

\item{y1var}{Is the name of the potential outcome to treatment}

\item{prop_blocks_0}{Is the proportion of blocks with no effects at all}

\item{tau_fn}{Is a function that turns ybase into the potential outcome under treatment --- it is a treatment effect creating function.}

\item{tau_size}{Is the parameter for the tau_fn --- like the true average effect size within a block.}

\item{pfn}{A function to produce pvalues --- using idat.}

\item{afn}{A function to adjust alpha at each step. Takes one or more p-values plus a stratum or batch indicator.}

\item{p_adj_method}{Must be "split" here.}

\item{copydts}{TRUE or FALSE. TRUE if using findBlocks standalone. FALSE if copied objects are being sent to findBlocks from other functions.}

\item{splitfn}{A function to split the data into two pieces --- using bdat}

\item{splitby}{A string indicating which column in bdat contains a variable to guide splitting (for example, a column with block sizes or block harmonic mean weights or a column with a covariate (or a function of covariates))}

\item{thealpha}{Is the error rate for a given test (for cases where alphafn is NULL, or the starting alpha for alphafn not null)}
}
\value{
False positive proportion out of the tests across the blocks, The false discovery rate (proportion rejected of false nulls out of all rejections), the power of the adjusted tests across blocks (the proportion of correctly rejected hypotheses out of all correct hypotheses --- in this case correct means non-null), and power of the unadjusted test (proportion correctly rejected out of  all correct hypothesis, but using unadjusted p-values).
}
\description{
Repeat experiment, reveal treatment effects from the potential outcomes, test within partitions, summarize
}
\details{
The function does hypothesis tests within partitions of blocks (including individual blocks depending on the splitting algorithmn) and then summarizes the results  of this testing across the blocks. It very much depends  on padj_test_fn.
}
