% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting_fns.R
\name{splitCluster}
\alias{splitCluster}
\title{Splitting function: K-Means Clustering}
\usage{
splitCluster(bid, x)
}
\arguments{
\item{bid}{Block id}

\item{x}{A vector that we can use to order the blocks}
}
\value{
A factor categorizing blocks into groups.
}
\description{
A splitting function takes block ids and a block ordering
vector (or vectors) and produces a factor that assigns some block ids to one
group or another group.
}
