% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting_fns.R
\name{splitEqualApprox}
\alias{splitEqualApprox}
\title{Splitting function: Approx Equal Splits}
\usage{
splitEqualApprox(bid, x)
}
\arguments{
\item{bid}{Block id}

\item{x}{A vector that we can use to order the blocks}
}
\value{
A factor categorizing blocks into groups.
}
\description{
A splitting function takes block ids and block ordering
vector (or vectors) and produces a factor that assigns some block ids to one
group or another group.
}
