% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitting_fns.R
\name{splitSpecifiedFactor}
\alias{splitSpecifiedFactor}
\title{A set of pre-specified splits}
\usage{
splitSpecifiedFactor(bid, x)
}
\arguments{
\item{x}{Is a a factor with levels like "state.district.school". The splits will occur from left to right depending on whether there is existing variation at that level}
}
\description{
A set of pre-specified splits
}
